Update README.md
Reminder that this code has not been updated for years and if I were to update it today I would use rewrite how I handled dark mode, improve component hierarchy, consider using a context or Zustand/Redux toolkit for global state and fix a few tiny UX issues. But since this is not a real project and just my portfolio, I think it looks great how it is!